+++ crates/node/src/main.rs
@@ -3,14 +3,15 @@
 #![warn(clippy::unwrap_used)]
 
 mod cli;
+mod extensions;
 mod service;
 mod util;
 
-use std::sync::Arc;
-
 use anyhow::Context;
 use clap::Parser;
+use extensions::madara_exexs;
 use mc_block_import::BlockImporter;
+use std::sync::Arc;
 
 use mc_db::DatabaseService;
 use mc_mempool::{GasPriceProvider, L1DataProvider, Mempool};
@@ -18,6 +19,7 @@ use mc_metrics::MetricsService;
 use mc_rpc::providers::{AddTransactionProvider, ForwardToProvider, MempoolAddTxProvider};
 use mc_telemetry::{SysInfo, TelemetryService};
 use mp_convert::ToFelt;
+use mp_exex::ExExLauncher;
 use mp_utils::service::{Service, ServiceGroup};
 
 use starknet_providers::SequencerGatewayProvider;
@@ -117,62 +119,69 @@ async fn main() -> anyhow::Result<()> {
 
     // Block provider startup.
     // `rpc_add_txs_method_provider` is a trait object that tells the RPC task where to put the transactions when using the Write endpoints.
-    let (block_provider_service, rpc_add_txs_method_provider): (_, Arc<dyn AddTransactionProvider>) =
-        match run_cmd.is_sequencer() {
-            // Block production service. (authority)
-            true => {
-                let mempool = Arc::new(Mempool::new(Arc::clone(db_service.backend()), Arc::clone(&l1_data_provider)));
+    let (block_provider_service, rpc_add_txs_method_provider): (_, Arc<dyn AddTransactionProvider>) = match run_cmd
+        .is_sequencer()
+    {
+        // Block production service. (authority)
+        true => {
+            let mempool = Arc::new(Mempool::new(Arc::clone(db_service.backend()), Arc::clone(&l1_data_provider)));
+            let mempool_provider = MempoolAddTxProvider::new(Arc::clone(&mempool));
 
-                let block_production_service = BlockProductionService::new(
-                    &run_cmd.block_production_params,
-                    &db_service,
-                    Arc::clone(&mempool),
-                    importer,
-                    Arc::clone(&l1_data_provider),
-                    run_cmd.devnet,
-                    prometheus_service.registry(),
-                    telemetry_service.new_handle(),
-                )?;
+            // Launch the ExEx manager for configured ExExs - if any.
+            let exex_manager = ExExLauncher::new(Arc::clone(&chain_config), madara_exexs()).launch().await?;
 
-                (ServiceGroup::default().with(block_production_service), Arc::new(MempoolAddTxProvider::new(mempool)))
-            }
-            // Block sync service. (full node)
-            false => {
-                // Feeder gateway sync service.
-                let sync_service = SyncService::new(
-                    &run_cmd.sync_params,
-                    Arc::clone(&chain_config),
-                    run_cmd.network.context(
-                        "You should provide a `--network` argument to ensure you're syncing from the right FGW",
-                    )?,
-                    &db_service,
-                    importer,
-                    telemetry_service.new_handle(),
-                )
-                .await
-                .context("Initializing sync service")?;
+            let block_production_service = BlockProductionService::new(
+                &run_cmd.block_production_params,
+                &db_service,
+                Arc::clone(&mempool),
+                importer,
+                Arc::clone(&l1_data_provider),
+                run_cmd.devnet,
+                exex_manager,
+                prometheus_service.registry(),
+                telemetry_service.new_handle(),
+            )?;
 
-                (
-                ServiceGroup::default().with(sync_service),
-                // TODO(rate-limit): we may get rate limited with this unconfigured provider?
-                Arc::new(ForwardToProvider::new(SequencerGatewayProvider::new(
-                    run_cmd
-                        .network
-                        .context(
-                            "You should provide a `--network` argument to ensure you're syncing from the right gateway",
-                        )?
-                        .gateway(),
-                    run_cmd
-                        .network
-                        .context(
-                            "You should provide a `--network` argument to ensure you're syncing from the right FGW",
-                        )?
-                        .feeder_gateway(),
-                    chain_config.chain_id.to_felt(),
-                ))),
+            (ServiceGroup::default().with(block_production_service), Arc::new(mempool_provider))
+        }
+        // Block sync service. (full node)
+        false => {
+            // TODO(rate-limit): we may get rate limited with this unconfigured provider?
+            let gateway_provider = ForwardToProvider::new(SequencerGatewayProvider::new(
+                run_cmd
+                    .network
+                    .context(
+                        "You should provide a `--network` argument to ensure you're syncing from the right gateway",
+                    )?
+                    .gateway(),
+                run_cmd
+                    .network
+                    .context("You should provide a `--network` argument to ensure you're syncing from the right FGW")?
+                    .feeder_gateway(),
+                chain_config.chain_id.to_felt(),
+            ));
+
+            // Launch the ExEx manager for configured ExExs - if any.
+            let exex_manager = ExExLauncher::new(Arc::clone(&chain_config), madara_exexs()).launch().await?;
+
+            // Feeder gateway sync service.
+            let sync_service = SyncService::new(
+                &run_cmd.sync_params,
+                Arc::clone(&chain_config),
+                run_cmd
+                    .network
+                    .context("You should provide a `--network` argument to ensure you're syncing from the right FGW")?,
+                &db_service,
+                importer,
+                exex_manager,
+                telemetry_service.new_handle(),
             )
-            }
-        };
+            .await
+            .context("Initializing sync service")?;
+
+            (ServiceGroup::default().with(sync_service), Arc::new(gateway_provider))
+        }
+    };
 
     let rpc_service = RpcService::new(
         &run_cmd.rpc_params,