+++ crates/primitives/exex/src/event.rs
@@ -1,0 +1,6 @@
+use starknet_api::block::BlockNumber;
+
+/// Events emitted by an `ExEx`.
+pub enum ExExEvent {
+    FinishedHeight(BlockNumber),
+}